Signing schemes test
